+++ cpp/src/snmg/degree/degree.cu
@@ -15,7 +15,8 @@
  */
 
 #include "degree.cuh"
-namespace cugraph {
+namespace cugraph { 
+namespace snmg {
 /**
  * Single node multi-GPU method for degree calculation on a partitioned graph.
  * @param x Indicates whether to compute in degree, out degree, or the sum of both.
@@ -57,11 +58,11 @@ gdf_error snmg_degree(int x, size_t* part_off, idx_t* off, idx_t* ind, idx_t** d
                     static_cast<idx_t>(env.get_num_sm() * 32));
     nblocks.y = 1;
     nblocks.z = 1;
-    degree_coo<idx_t, idx_t> <<<nblocks, nthreads>>>(static_cast<idx_t>(loc_e),
+    cugraph::detail::degree_coo<idx_t, idx_t> <<<nblocks, nthreads>>>(static_cast<idx_t>(loc_e),
                                                      static_cast<idx_t>(loc_e),
                                                      ind,
                                                      local_result);
-    cudaCheckError();
+    CUDA_CHECK_LAST();
   }
 
   // Out-degree
@@ -74,11 +75,11 @@ gdf_error snmg_degree(int x, size_t* part_off, idx_t* off, idx_t* ind, idx_t** d
                     static_cast<idx_t>(env.get_num_sm() * 32));
     nblocks.y = 1;
     nblocks.z = 1;
-    degree_offsets<idx_t, idx_t> <<<nblocks, nthreads>>>(static_cast<idx_t>(loc_v),
+    cugraph::detail::degree_offsets<idx_t, idx_t> <<<nblocks, nthreads>>>(static_cast<idx_t>(loc_v),
                                                          static_cast<idx_t>(loc_e),
                                                          off,
                                                          local_result + part_off[i]);
-    cudaCheckError();
+    CUDA_CHECK_LAST();
   }
 
   // Combining the local results into global results
@@ -126,11 +127,11 @@ gdf_error snmg_degree<int64_t>(int x,
                     static_cast<int64_t>(env.get_num_sm() * 32));
     nblocks.y = 1;
     nblocks.z = 1;
-    degree_coo<int64_t, double> <<<nblocks, nthreads>>>(static_cast<int64_t>(loc_e),
+    cugraph::detail::degree_coo<int64_t, double> <<<nblocks, nthreads>>>(static_cast<int64_t>(loc_e),
                                                         static_cast<int64_t>(loc_e),
                                                         ind,
                                                         reinterpret_cast<double*>(local_result));
-    cudaCheckError();
+    CUDA_CHECK_LAST();
   }
 
   // Out-degree
@@ -143,12 +144,12 @@ gdf_error snmg_degree<int64_t>(int x,
                     static_cast<int64_t>(env.get_num_sm() * 32));
     nblocks.y = 1;
     nblocks.z = 1;
-    degree_offsets<int64_t, double> <<<nblocks, nthreads>>>(static_cast<int64_t>(loc_v),
+    cugraph::detail::degree_offsets<int64_t, double> <<<nblocks, nthreads>>>(static_cast<int64_t>(loc_v),
                                                             static_cast<int64_t>(loc_e),
                                                             off,
                                                             reinterpret_cast<double*>(local_result
                                                                 + part_off[i]));
-    cudaCheckError();
+    CUDA_CHECK_LAST();
   }
 
   // Convert the values written as doubles back to int64:
@@ -160,8 +161,8 @@ gdf_error snmg_degree<int64_t>(int x,
                   static_cast<int64_t>(env.get_num_sm() * 32));
   nblocks.y = 1;
   nblocks.z = 1;
-  type_convert<double, int64_t> <<<nblocks, nthreads>>>(reinterpret_cast<double*>(local_result), glob_v);
-  cudaCheckError();
+  cugraph::detail::type_convert<double, int64_t> <<<nblocks, nthreads>>>(reinterpret_cast<double*>(local_result), glob_v);
+  CUDA_CHECK_LAST();
 
   // Combining the local results into global results
   treeReduce<int64_t, thrust::plus<int64_t> >(env, glob_v, local_result, degree);
@@ -172,7 +173,7 @@ gdf_error snmg_degree<int64_t>(int x,
   return GDF_SUCCESS;
 }
 
-}//namespace
+} } //namespace
 
 
 template<typename idx_t>
@@ -181,22 +182,22 @@ gdf_error gdf_snmg_degree_impl(int x,
                                gdf_column* off,
                                gdf_column* ind,
                                gdf_column** x_cols) {
-  GDF_REQUIRE(off->size > 0, GDF_INVALID_API_CALL);
-  GDF_REQUIRE(ind->size > 0, GDF_INVALID_API_CALL);
-  GDF_REQUIRE(off->dtype == ind->dtype, GDF_UNSUPPORTED_DTYPE);
-  GDF_REQUIRE(off->null_count + ind->null_count == 0, GDF_VALIDITY_UNSUPPORTED);
+  CUGRAPH_EXPECTS(off->size > 0, "Invalid API parameter");
+  CUGRAPH_EXPECTS(ind->size > 0, "Invalid API parameter");
+  CUGRAPH_EXPECTS(off->dtype == ind->dtype, "Unsupported data type");
+  CUGRAPH_EXPECTS(off->null_count + ind->null_count == 0, "Column must be valid");
 
   gdf_error status;
   auto p = omp_get_num_threads();
 
   idx_t* degree[p];
   for (auto i = 0; i < p; ++i) {
-    GDF_REQUIRE(x_cols[i] != nullptr, GDF_INVALID_API_CALL);
-    GDF_REQUIRE(x_cols[i]->size > 0, GDF_INVALID_API_CALL);
+    CUGRAPH_EXPECTS(x_cols[i] != nullptr, "Invalid API parameter");
+    CUGRAPH_EXPECTS(x_cols[i]->size > 0, "Invalid API parameter");
     degree[i] = static_cast<idx_t*>(x_cols[i]->data);
   }
 
-  status = cugraph::snmg_degree(x,
+  status = cugraph::snmg::snmg_degree(x,
                                 part_offsets,
                                 static_cast<idx_t*>(off->data),
                                 static_cast<idx_t*>(ind->data),
@@ -209,10 +210,10 @@ gdf_error gdf_snmg_degree(int x,
                           gdf_column* off,
                           gdf_column* ind,
                           gdf_column** x_cols) {
-  GDF_REQUIRE(part_offsets != nullptr, GDF_INVALID_API_CALL);
-  GDF_REQUIRE(off != nullptr, GDF_INVALID_API_CALL);
-  GDF_REQUIRE(ind != nullptr, GDF_INVALID_API_CALL);
-  GDF_REQUIRE(x_cols != nullptr, GDF_INVALID_API_CALL);
+  CUGRAPH_EXPECTS(part_offsets != nullptr, "Invalid API parameter");
+  CUGRAPH_EXPECTS(off != nullptr, "Invalid API parameter");
+  CUGRAPH_EXPECTS(ind != nullptr, "Invalid API parameter");
+  CUGRAPH_EXPECTS(x_cols != nullptr, "Invalid API parameter");
   switch (off->dtype) {
     case GDF_INT32:
       return gdf_snmg_degree_impl<int32_t>(x, part_offsets, off, ind, x_cols);