+++ cpp/src/traversal/sssp.cu
@@ -25,7 +25,8 @@
 #include "sssp_kernels.cuh"
 #include "utilities/error_utils.h"
 
-namespace cugraph {
+namespace cugraph { 
+namespace detail {
 
 template <typename IndexType, typename DistType>
 void SSSP<IndexType, DistType>::setup() {
@@ -233,7 +234,7 @@ gdf_error SSSP<IndexType, DistType>::traverse(IndexType source_vertex) {
                     cudaMemcpyDeviceToDevice,
                     stream);
 
-    cudaCheckError();
+    CUDA_CHECK_LAST();
 
     // We need nf for the loop
     cudaStreamSynchronize(stream);
@@ -275,7 +276,7 @@ void SSSP<IndexType, DistType>::clean() {
   ALLOC_FREE_TRY(next_distances, nullptr);
 }
 
-}  // end namespace cugraph
+} } //namespace
 
 /**
  * ---------------------------------------------------------------------------*
@@ -289,7 +290,7 @@ gdf_error gdf_sssp(gdf_graph* gdf_G,
                    gdf_column* predecessors,
                    const int source_vert) {
 
-  GDF_REQUIRE(gdf_G->adjList != nullptr, GDF_INVALID_API_CALL);
+  CUGRAPH_EXPECTS(gdf_G->adjList != nullptr, "Invalid API parameter");
 
   void *sssp_dist_ptr, *pred_ptr;
   // NOTE: gdf_column struct doesn't have a default constructor. So we can get
@@ -302,28 +303,29 @@ gdf_error gdf_sssp(gdf_graph* gdf_G,
   pred_ptr =
       (predecessors && predecessors->size) ? predecessors->data : nullptr;
 
-  GDF_REQUIRE(sssp_dist_ptr || pred_ptr, GDF_INVALID_API_CALL);
+  CUGRAPH_EXPECTS(sssp_dist_ptr || pred_ptr, "Invalid API parameter");
 
   if (sssp_dist_ptr) {
-    GDF_REQUIRE(!sssp_distances->valid, GDF_VALIDITY_UNSUPPORTED);
+    CUGRAPH_EXPECTS(!sssp_distances->valid, "Column must be valid");
     // Integral types are possible, but we don't want to deal with overflow
     // conditions right now
-    GDF_REQUIRE(sssp_distances->dtype == GDF_FLOAT32 ||
+    CUGRAPH_EXPECTS(sssp_distances->dtype == GDF_FLOAT32 ||
                     sssp_distances->dtype == GDF_FLOAT64,
-                GDF_INVALID_API_CALL);
+                "Invalid API parameter");
   }
-  GDF_REQUIRE(gdf_G->adjList->offsets->dtype == GDF_INT32,
-              GDF_UNSUPPORTED_DTYPE);
-  GDF_REQUIRE(gdf_G->adjList->indices->dtype == GDF_INT32,
-              GDF_UNSUPPORTED_DTYPE);
+
+  CUGRAPH_EXPECTS(gdf_G->adjList->offsets->dtype == GDF_INT32,
+              "Unsupported data type");
+  CUGRAPH_EXPECTS(gdf_G->adjList->indices->dtype == GDF_INT32,
+              "Unsupported data type");
 
   if (pred_ptr)
-    GDF_REQUIRE(predecessors->dtype == gdf_G->adjList->indices->dtype,
-                GDF_UNSUPPORTED_DTYPE);
+    CUGRAPH_EXPECTS(predecessors->dtype == gdf_G->adjList->indices->dtype,
+                "Unsupported data type");
 
   if (sssp_dist_ptr)
-    GDF_REQUIRE(gdf_G->adjList->offsets->size - 1 <= sssp_distances->size,
-                GDF_INVALID_API_CALL);
+    CUGRAPH_EXPECTS(gdf_G->adjList->offsets->size - 1 <= sssp_distances->size,
+                "Invalid API parameter");
 
   if (!gdf_G->adjList->edge_data) {
     // Generate unit weights
@@ -371,17 +373,17 @@ gdf_error gdf_sssp(gdf_graph* gdf_G,
     }
   } else {
     // Got weighted graph
-    GDF_REQUIRE(
+    CUGRAPH_EXPECTS(
         gdf_G->adjList->edge_data->size == gdf_G->adjList->indices->size,
-        GDF_INVALID_API_CALL);
+        "Invalid API parameter");
 
-    GDF_REQUIRE(gdf_G->adjList->edge_data->dtype == GDF_FLOAT32 ||
+    CUGRAPH_EXPECTS(gdf_G->adjList->edge_data->dtype == GDF_FLOAT32 ||
                     gdf_G->adjList->edge_data->dtype == GDF_FLOAT64,
-                GDF_INVALID_API_CALL);
+                "Invalid API parameter");
 
     if (sssp_dist_ptr)
-      GDF_REQUIRE(gdf_G->adjList->edge_data->dtype == sssp_distances->dtype,
-                  GDF_UNSUPPORTED_DTYPE);
+      CUGRAPH_EXPECTS(gdf_G->adjList->edge_data->dtype == sssp_distances->dtype,
+                  "Unsupported data type");
 
     // SSSP is not defined for graphs with negative weight cycles
     // Warn user about any negative edges
@@ -399,7 +401,7 @@ gdf_error gdf_sssp(gdf_graph* gdf_G,
   gdf_error ret;
 
   if (gdf_G->adjList->edge_data->dtype == GDF_FLOAT32) {
-    cugraph::SSSP<int, float> sssp(
+    cugraph::detail::SSSP<int, float> sssp(
         n, e, offsets_ptr, indices_ptr, static_cast<float*>(edge_weights_ptr));
 
     sssp.configure(static_cast<float*>(sssp_dist_ptr),
@@ -408,7 +410,7 @@ gdf_error gdf_sssp(gdf_graph* gdf_G,
 
     ret = sssp.traverse(source_vert);
   } else if (gdf_G->adjList->edge_data->dtype == GDF_FLOAT64) {
-    cugraph::SSSP<int, double> sssp(n,
+    cugraph::detail::SSSP<int, double> sssp(n,
                                     e,
                                     offsets_ptr,
                                     indices_ptr,
@@ -420,9 +422,9 @@ gdf_error gdf_sssp(gdf_graph* gdf_G,
 
     ret = sssp.traverse(source_vert);
   } else {
-    GDF_REQUIRE(gdf_G->adjList->edge_data->dtype == GDF_FLOAT32 ||
+    CUGRAPH_EXPECTS(gdf_G->adjList->edge_data->dtype == GDF_FLOAT32 ||
                     gdf_G->adjList->edge_data->dtype == GDF_FLOAT64,
-                GDF_INVALID_API_CALL);
+                "Invalid API parameter");
   }
 
   return ret;